CSC 113 Final Project - Mahian Khan
This project is a Python script that automates the process of checking the availability of certain products on the Best Buy website and sending notifications to a Discord channel when the availability of a product changes. The script uses a checkStock class to organize the relevant functionality and includes techniques such as the use of HTTP and HTTPS proxies, multithreading, and a queue to optimize its performance. The script provides a convenient and efficient way to track the availability of specific products on the Best Buy website and receive notifications when the availability changes.